Complete time count and save

+++ app/src/main/java/com/example/timemanager/room_components/TimeDataEntity.kt
@@ -7,6 +7,6 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "TDE")
 data class TimeDataEntity(
     @PrimaryKey(autoGenerate = true) val id: Int = 0,
-    @ColumnInfo("timeData") val timeData: Long,
+    @ColumnInfo("timeData") val timeData: Int,
     @ColumnInfo("doing") val doing: String
 )
